refactor: GuestGNB 리팩토링
