+++ src/components/common/GuestGNB/style.ts
@@ -28,29 +28,14 @@ export const Inner = styled.header`
   display: flex;
   align-items: center;
   max-width: 1440px;
-  height: 4.5rem;
+  height: 72px;
   margin: 0 auto;
-  padding: 0 2rem;
+  padding: 0 32px;
 
   @media ${({ theme }) => theme.mediaQuery.md} {
+    height: 60px;
     padding: 0 8px;
   }
-`;
-
-export const Logo = styled.div`
-  display: flex;
-  gap: 0.4rem;
-
-  margin-right: 3rem;
-
-  @media ${({ theme }) => theme.mediaQuery.md} {
-    margin-right: 0;
-  }
-`;
-
-export const LogoText = styled.span`
-  font-size: 1.5rem;
-  font-weight: 600;
 `;
 
 export const NavArea = styled.nav`
@@ -96,50 +81,6 @@ export const NavItem = styled(Link)<{ active: number }>`
     margin-right: 32px;
     height: auto;
   }
-`;
-
-export const ActionArea = styled.div`
-  display: flex;
-`;
-
-export const TILInfo = styled.div`
-  background-color: ${({ theme }) => theme.colors.gray_200};
-  padding: 1.125rem;
-  border-radius: 1rem;
-
-  & > span:nth-of-type(2) {
-    color: ${({ theme }) => theme.colors.rose};
-  }
-`;
-
-export const TILButtonStyles = (theme: EmotionTheme) => css`
-  font-size: 1.25rem;
-  font-weight: 700;
-
-  @media ${theme.mediaQuery.md} {
-    margin-right: 8px;
-  }
-`;
-
-export const NoticeContainer = styled.div`
-  position: relative;
-  display: flex;
-  align-items: center;
-`;
-
-export const AlarmActiveDot = styled.div`
-  position: absolute;
-  top: 6px;
-  right: -4px;
-  width: 1rem;
-  height: 1rem;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.colors.rose};
-`;
-
-export const ProfileSkeletonStyles = css`
-  width: 40px;
-  height: 40px;
 `;
 
 export const ButtonStyles = (theme: EmotionTheme) => css`
